Move FileSystemSnapshotBuilder to api package
