+++ subprojects/snapshots/src/main/java/org/gradle/internal/snapshot/FileSystemSnapshotter.java
@@ -18,7 +18,6 @@ package org.gradle.internal.snapshot;
 
 import org.gradle.api.tasks.util.PatternSet;
 import org.gradle.internal.hash.HashCode;
-import org.gradle.internal.snapshot.impl.FileSystemSnapshotBuilder;
 
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.ThreadSafe;
